Fixed data format

+++ AutoTrashCartWebServiceApp/Models/Location.cs
@@ -4,11 +4,11 @@
     {
         public string Token { get; set; }
         public string LocationType { get; set; }
-        public float Latitude0 { get; set; }
-        public float Longitude0 { get; set; }
-        public float Latitude1 { get; set; }
-        public float Longitude1 { get; set; }
-        public float Latitude2 { get; set; }
-        public float Longitude2 { get; set; }
+        public decimal Latitude0 { get; set; }
+        public decimal Longitude0 { get; set; }
+        public decimal Latitude1 { get; set; }
+        public decimal Longitude1 { get; set; }
+        public decimal Latitude2 { get; set; }
+        public decimal Longitude2 { get; set; }
     }
 }
