+++ AutoTrashCartWebServiceApp/Models/Orientation.cs
@@ -4,8 +4,8 @@
     {
         public string Token { get; set; }
         public string OrientationType { get; set; }
-        public float X { get; set; }
-        public float Y { get; set; }
-        public float Z { get; set; }
+        public decimal X { get; set; }
+        public decimal Y { get; set; }
+        public decimal Z { get; set; }
     }
 }
